started on refactoring, added SqlIntegrator class
