+++ Postulate.Lite.SqlServer/IntKey/ConnectionExtensions.cs
@@ -10,7 +10,7 @@ namespace Postulate.Lite.SqlServer.IntKey
 	{
 		private static SqlServerProvider<int> GetProvider()
 		{
-			return new SqlServerProvider<int>((obj) => Convert.ToInt32(obj), "identity(1,1)");
+			return new SqlServerProvider<int>((obj) => Convert.ToInt32(obj));
 		}
 
 		public static bool Exists<TModel>(this IDbConnection connection, int id, IUser user = null)
