+++ Postulate.Lite.SqlServer/Postulate.Lite.SqlServer.csproj
@@ -49,6 +49,7 @@
     <Compile Include="GuidKey\ConnectionExtensions.cs" />
     <Compile Include="IntKey\ConnectionExtensions.cs" />
     <Compile Include="LongKey\ConnectionExtensions.cs" />
+    <Compile Include="SqlServerIntegrator.cs" />
     <Compile Include="SqlServerProvider_Crud.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SqlServerProvider_Merge.cs" />
