+++ Postulate.Lite.SqlServer/SqlServerIntegrator.cs
@@ -1,0 +1,33 @@
+﻿using Postulate.Lite.Core;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Postulate.Lite.SqlServer
+{
+	public class SqlServerIntegrator : SqlIntegrator
+	{
+		public override Dictionary<Type, string> SupportedTypes(int length, int precision, int scale)
+		{
+			return new Dictionary<Type, string>()
+			{
+				{ typeof(string), $"nvarchar({((length == 0) ? "max" : length.ToString())})" },
+				{ typeof(int), "int" },
+				{ typeof(DateTime), "datetime" },
+				{ typeof(bool), "bit" },
+				{ typeof(decimal), $"decimal({scale}, {precision})" },
+				{ typeof(long), $"bigint" },
+				{ typeof(short), "smallint" },
+				{ typeof(byte), "tinyint" },
+				{ typeof(TimeSpan), "time" },
+				{ typeof(double), "float" },
+				{ typeof(float), "float" },
+				{ typeof(Guid), "uniqueidentifier" },
+				{ typeof(char), "char(1)" },
+				{ typeof(byte[]), $"varbinary({length})" }
+			};
+		}
+	}
+}
