+++ Postulate.Lite.SqlServer/SqlServerProvider_Crud.cs
@@ -11,7 +11,7 @@ namespace Postulate.Lite.SqlServer
 {
 	public partial class SqlServerProvider<TKey> : CommandProvider<TKey>
 	{
-		public SqlServerProvider(Func<object, TKey> identityConverter, string identitySyntax) : base(identityConverter, identitySyntax)
+		public SqlServerProvider(Func<object, TKey> identityConverter) : base(identityConverter, new SqlServerIntegrator())
 		{
 		}
 
@@ -23,14 +23,14 @@ namespace Postulate.Lite.SqlServer
 		protected override string FindCommand<T>(string whereClause)
 		{
 			var type = typeof(T);
-			var props = GetMappedColumns(type);
+			var props = _integrator.GetMappedColumns(type);
 			var columns = props.Select(pi => new ColumnInfo(pi));
 			return $"SELECT {string.Join(", ", columns.Select(col => ApplyDelimiter(col.ColumnName)))} FROM {ApplyDelimiter(TableName(typeof(T)))} WHERE {whereClause}";
 		}
 
 		protected override string InsertCommand<T>()
 		{
-			var columns = EditableColumns<T>(SaveAction.Insert);
+			var columns = _integrator.EditableColumns<T>(SaveAction.Insert);
 			string columnList = string.Join(", ", columns.Select(c => ApplyDelimiter(c.ColumnName)));
 			string valueList = string.Join(", ", columns.Select(c => $"@{c.PropertyName}"));
 			return $"INSERT INTO {ApplyDelimiter(TableName(typeof(T)))} ({columnList}) OUTPUT [inserted].[{typeof(T).GetIdentityName()}] VALUES ({valueList});";
@@ -38,35 +38,14 @@ namespace Postulate.Lite.SqlServer
 
 		protected override string UpdateCommand<T>()
 		{
-			var columns = EditableColumns<T>(SaveAction.Update);
+			var columns = _integrator.EditableColumns<T>(SaveAction.Update);
 			return $"UPDATE {ApplyDelimiter(TableName(typeof(T)))} SET {string.Join(", ", columns.Select(col => $"{ApplyDelimiter(col.ColumnName)}=@{col.PropertyName}"))} WHERE {ApplyDelimiter(typeof(T).GetIdentityName())}=@id";
 		}
 
 		protected override string DeleteCommand<T>()
 		{
 			return $"DELETE {ApplyDelimiter(TableName(typeof(T)))} WHERE {ApplyDelimiter(typeof(T).GetIdentityName())}=@id";
-		}
-
-		protected override Dictionary<Type, string> SupportedTypes(int length, int precision, int scale)
-		{
-			return new Dictionary<Type, string>()
-			{
-				{ typeof(string), $"nvarchar({((length == 0) ? "max" : length.ToString())})" },
-				{ typeof(int), "int" },
-				{ typeof(DateTime), "datetime" },
-				{ typeof(bool), "bit" },
-				{ typeof(decimal), $"decimal({scale}, {precision})" },
-				{ typeof(long), $"bigint" },
-				{ typeof(short), "smallint" },
-				{ typeof(byte), "tinyint" },
-				{ typeof(TimeSpan), "time" },
-				{ typeof(double), "float" },
-				{ typeof(float), "float" },
-				{ typeof(Guid), "uniqueidentifier" },
-				{ typeof(char), "char(1)" },
-				{ typeof(byte[]), $"varbinary({length})" }
-			};
-		}
+		}		
 
 		protected override string TableName(Type modelType)
 		{
